# Use sets instead of arrays for better performance.