This further improves the running time on my machine
from 3.8s down to 2.3s for the default (Everest) board,
and from 10s down to 6s for the "difficulty: off charts :)"
example from collections/test_db.txt; that is:
  python2 sudoku.py 1.....7....71.9...68..7......1.9.6.....3...2..4......3..8.6.1..5......4......2..5

Also changed the logic for creating the groups to be clearer (to me, anyway).

====================================================
Note: The above description is from the third commit e369810 of this pull request; the first two commits 80fd632 and  d04bd94 is are previous pull request #2 and pull request #3 which should be pulled first (if acceptable). I'm not sure I'm managing these dependent pull requests properly.