@r1cc4rdo 2021-04-20
Will take a longer look at this one.
Curious to know if the performance savings are mostly due to groups' sets creation (which in my code happens over and over). The general rule of thumb says that lists are often faster when the number of elements is low. Good catch :)

Might take a few days.