@donhatch 2021-04-20
If I recall correctly there are two tweaks in this commit:

- moving the repeated conversion from list to set (of both `subset`, and of `g` from `groups`) out of the loop
- changing the groups to be sets in the first place, so that no repeated conversion of those necessary at all.

I observed that each of these tweaks caused a measurable improvement.  (What I described in the commit message was the combined improvement of both.)

But you're right, it's possible that changing groups to be sets may have hurt performance in other parts of the code, even though it helped on balance; I didn't look into that possibility.