@r1cc4rdo 2021-04-20
Yeah, I trust your conclusions -- just want to see it for myself before taking it in.