@r1cc4rdo 2021-04-27
I've reviewed the runtime for this and several variations (including one that moves most of the eliminate_candidates logic during initialization, attached below). Can confirm that the source of inefficiency was the repeated set creation.

I'm going to pull this for the time being, and consider if to amend in the future with the version below, which is currently the fastest without adding additional logic.

```
col_to_indexes = [list(range(i, 81, 9)) for i in range(9)]
row_to_indexes = [list(range(i, i + 9)) for i in range(0, 81, 9)]
block_to_indexes = [[c + 9*r + 3*bc + 27*br for r in range(3) for c in range(3)]
                    for br in range(3) for bc in range(3)]  # block row and col
groups = tuple(map(set, row_to_indexes + col_to_indexes + block_to_indexes))
subsets = sorted({subset for group in groups for subset_size in range(1, 9)
                  for subset in combinations(group, subset_size)}, key=lambda x: (len(x),) + x)
dependencies = {subset: {i for group in groups for i in group if set(subset) <= group and i not in subset}
                for subset in subsets}  # if we find a constraint in key cells, we need to update value cells


def eliminate_candidates(board, dependencies=dependencies):
    for subset, peers in dependencies.items():
        candidates_in_subset = set(''.join(board[index] for index in subset))
        if len(candidates_in_subset) == len(subset):  # we found a constraint
            for index in peers:
                board[index] = ''.join(c for c in board[index] if c not in candidates_in_subset)
                assert board[index]  # if triggered, inconsistent assignment detected
```